Supports RFC 5425 and RFC 6587 msg-len fields
In order to support long messages, and messages containing the full range of UTF-8 messages a message length may be conveyed at the beginning of a message. As defined by the following RFCs:

  https://tools.ietf.org/html/rfc5425
  https://tools.ietf.org/html/rfc6587

The enhancement changes the behaviour of fluentd's TcpHandler to accept a function that determines whether the tcp message has been completely read. In the absence of a message length field in the message, the original behaviour of using a delimiter ("\n") is used. Otherwise the message length is honoured.