Reject chunks without timekey from candidates of destination of symlink
"symlink" feature is to link from symlink_path to the latest file chunk. Records with latest timekey will be appended into that file chunk.
On the other side, resumed file chunks might NOT have timekey, especially in the cases that resumed file chunks are generated by
Fluentd v0.12. These chunks will be enqueued immediately, and will be flushed soon.

So these files SHOULD NOT be a candidate of destination of symlink.