Handle windows files paths in the in_tail plugin.
Currently tail plugins on windows will work if a specific file is targeted:
<source>
  type tail
  format none
  path C:\SomeDir\Logs\the.logs
  tag custom-logs
</source>

However with wildcards the log files will not be picked up:
<source>
  type tail
  format none
  path C:\SomeDir\Logs\**\*'
  tag custom-logs
</source>

This is due to Dir.glob not handing '\'.  The '\' must be converted to '/'.  FIle.* methods will take either a '\' or '/' so just changing all '\' from the start is the easist.

Users could define paths with '/'s but this is not idomatic for windows users.