in_http: Do not discard the 'time' field in incoming JSON records.
It was found that parse_params_default() unconditionally discards the
`time` value returned by the JSON parser. This trivially leads to
silent data loss.

We can confirm this by sending a HTTP request like below:

  $ curl -X POST -d 'json={"time":1}' http://localhost:8888/debug.log

It results in the following line, which loses the original JSON data
completely:

  2018-02-21 11:45:30.124988214 +0900 debug.log: {}

This patch fixes the issue by passing the time stamp (extracted from
the incoming JSON data) to the caller.