Edit Fluentd license info so GitHub recognizes it
GitHub uses a library called Licensee to identify a project's license
type. It shows this information in the status bar and via the API if it
can unambiguously identify the license.

This commit removes the COPYING file (which duplicates info that's
already included in LICENSE), so that the license type identification
(as Apache 2.0) will succeed.

Signed-off-by: Andrea Kao <eirinikos@gmail.com>