test_retry_state: Fix a test failure on AppVeyor with Ruby 2.3 (x86)
This fixes 'check inf handling' test not to fail on AppVeyor CI anymore.

Specifically, in this test script, it is assumed that `0.1 * 2 ** 10.25`
overflows the float boundary and becomes an infinite number. This is not
quite true because the number is roughly 1.79e+307 and it's actually well
below the upper boundary of double-precision floats.

Still, even though the resulting number is within the maximum limit, the
test case passes on *most* environments because an overflow occurs
while computing raw_interval():

    def raw_interval(num)
      @constant_factor.to_f * (@backoff_base ** (num))
    end

But evidently this is not reliable among environments. So let's make
the interval actually go over the boundary, instead of relying on an
unreliable behaviour.

Signed-off-by: Fujimoto Seiji <fujimoto@clear-code.com>