in_http: Fix "ignored parser options" problem on bulk insertion
What used to happen is that options in the <parser> section get ignored
when we send events as a JSON array. This occurs because parser plugins
are designed to handle a single hash value and do not generallly handle
an array input well.

Until now, we have tried to solve this issue by emulating the semantics
of parser plugins in Fluent::Plugin::HttpInput (see 1afbfb1, 39f3a0d
and f560017c). However, this approach turned out to be error prone and
rather tedious.

This patch takes a different approach:

 - Whenever @parser is available, reuse @parser.convert_values() by
   manually applying it to each record in an input array.
 - Otherwise, fall back to the current logic.

... and should solve this class of issues.

Signed-off-by: Fujimoto Seiji <fujimoto@clear-code.com>