parser_json: Add stream_buffer_size config param
Allow configuration of the size of the buffer that Yajl uses when
parsing streaming input.

The advantage of this is that when using `out_exec_filter`, and parsing
as JSON, it's now possible to configure this plugin to avoid having to
wait for 8192 bytes of data to be parsed before events are emitted.

Configuration in the `out_exec_filter` tests has been modified to use
this parameter, as it shaves 60 seconds off the test run time.

Signed-off-by: Ben Wheatley <contact@benwh.com>