out_forward: do not use SO_LINGER on SSL/TLS WinSock
There is an obscure bug in out_forward that prevents events from
being sent to a remote node. It occures only on Windows, and is
reproducable only when SSL/TLS is enabled.

The root cause is not fully investigated yet, but disabling
SO_LINGER is confirmed to fix the issue. So this implements it.

Signed-off-by: Fujimoto Seiji <fujimoto@clear-code.com>