in_tail: Dump a file name in addition to a malformed line
This patch is based on a request from one of our customers.

Right now, in_tail emits the following warning message when it reads
a malformed line.

    2019-08-18 00:20:03 +0900 [warn]: #0 "a,b\",c" error="Illegal quoting in line 1."

However, since we use in_tail extensively (e.g. tracking lots of files
at once), it is sometimes not easy to determine which file caused the
error just by seeing this message.

With this, in_tail shows:

    2019-08-18 00:27:42 +0900 [warn]: #0 invalid line found file="test.csv"
    line="a,b\",c" error="Illegal quoting in line 1."

... which should be useful for debugging.

Signed-off-by: Fujimoto Seiji <fujimoto@ceptord.net>