out_http.rb: add more retryable_response_codes
We were surprised to find that the HTTP output
plugin by default considers a 500 HTTP response
to be unrecoverable.

In the real world most 5xx errors represent transient
problems.

In addition to retrying upon 503, also retrying upon
500, 502, and 504 will -- in most environments --
be the better default, leading to significantly less
_false negative_ "unrecoverable errors".

When a sane HTTP reverse proxy is involved
(example: nginx) then definitely retry upon 502
and 504:

- most of the times when nginx returns a 502 response
then it is actually a transient TCP error in the interaction
with the upstream server. Example error message
from nginx:

   [error] 2908#0: *247 recv() failed (104: Connection reset by peer)
   while reading response header from upstream

This is often transient and should be retried.

- most of the times when nginx returns a 504 response
then it is -- again -- actually a transient TCP error
in the interaction with the upstream server, such as
a TCP connection timeout.

For other HTTP servers that are not as established
and production-hardened as nginx, the 500 HTTP
error is often emitted as a catch-all to represent all
kinds of transient problems.

Signed-off-by: Dr. Jan-Philip Gehrcke <jp@opstrace.com>