Add fluent-ctl command
It sends a simple command to a fluentd process.
The main purpose of this command is that providing a kill command
alternative on Windows. On UNIX, it just only send a UNIX signal to a
specified process.

Usage: fluent-ctl COMMAND PID_OR_SIGNAME

Commands:
  shutdown
  restart
  flush
  reload

SIGNAME can be specified by fluentd's "-x" or "--signame" option.
Note that this command doesn't support sending events to windows
service.

Signed-off-by: Takuro Ashie <ashie@clear-code.com>