github: add a guide to google groups
In the previous version, there are only two choices for
bug and feature request.

So Fluentd user tends to use "Bug Report" even though it is not a
bug report definitely.

This PR try to add third choice explicitly to guide to
user forum.

Before:

  * Bug Report
  * Feature request

After:

  * Bug Report
  * Feature request
  * Ask a question

Additionally, the explanation about bug report is updated to emphasize
it is a "Bug Report".

Signed-off-by: Kentaro Hayashi <hayashi@clear-code.com>