Follow depended gem version by fluent-plugin-generate
In the previous versions, generated .gemspec doesn't follow
depended gem version which is used in fluentd.gemspec at all.

Instead, it is changed to retrieve depended gem version from
Gemfile.lock to sync with it.

Before:

  spec.add_development_dependency "bundler", "~> 1.14"
  spec.add_development_dependency "rake", "~> 12.0"
  spec.add_development_dependency "test-unit", "~> 3.0"

It is required to update manually.

After:

  spec.add_development_dependency "bundler", "~> 2.2.15"
  spec.add_development_dependency "rake", "~> 13.0"
  spec.add_development_dependency "test-unit", "~> 3.3"

Above depended version is replaced by <%= %> in template.
No need to update manually.

Signed-off-by: Kentaro Hayashi <kenhys@gmail.com>