Enable system.log.rotate_... system configuration
In the previous versions, log rotation options were supported as
--log-rotate-age or --log-rotate-size via command line options.

On Windows, as fluentd is launched as a windows service, it is
required to configure again via --reg-winsvc-fluentdopt or edit
fluentdopt registry key for customization.

This approach is not convenient for Windows users, so it may be better
to support more comprehensive solution - customize via configuration
file.

This commit introduces such a configurable parameter in .conf

  <system>
   <log>
     rotate_age 5
     rotate_size 1048576
   </log>
  </system>

Signed-off-by: Kentaro Hayashi <hayashi@clear-code.com>