test: declare that test case use v0.10 explicitly
In the previous versions, Fluent::Supervisor.new(opts) doesn't parse
content actually at that time, so it doesn't care whether v0 or v1
parser is used.

But, by introducing support for log rotation configuration, it was
changed to fetch rotate_age and rotate_size in early stage because
these parameters must be passed to LogInitializer during
initialization steps, so configuration file is changed to parse in
Fluent::Supervisor.new(opts).

As a result, we need to pass additional parameter for this test case.
Without this fix, log rotate configuration in <system> support raises
the following exception:

  Failure: test_inline(ConfigTest):
    Exception raised:
    Fluent::ConfigParseError(<expected end tag '</include>' but got end of file at config_test_1.conf line 11,10
     10:   port 2222
     11:  </source>
         ----------^
    >)

Signed-off-by: Kentaro Hayashi <hayashi@clear-code.com>