fluent-cat: support secondary file
https://docs.fluentd.org/output/secondary_file#how-to-resend-secondary-file
explains that it supports secondary file to resend,
but it get error when [Fluent::EventTime, Hash] format is given.

  Input must be a map (got Array)

In this commit, check record content and support [Fluent::EventTime,
Hash] format too.

Note that fluent-cat ignore timestamp which is stored in secondary
records. (it is intended behavior for keeping consistency)

Signed-off-by: Kentaro Hayashi <hayashi@clear-code.com>