in_tail: Shutdown immediately & safely even if reading huge files
in_tail takes very long time to shutdown when it's reading huge files.
This patch ensure to shutdown ASAP just after receiving a shutdown
command.

* Constructed tail watchers should be registered ASAP to be recognized
  by the shutdown process.
* Break TailWatcher#on_notify ASAP
  Wait 5 secdonds before breaking it to ensure reading remaining lines.

It takes effect only when `skip_refresh_on_startup` is enabled, because
reading files is occurred on main thread without it. It should be
enabled by default in the future.

Signed-off-by: Takuro Ashie <ashie@clear-code.com>