buffer: minimize data lost by overflow exception for compressed data
In the previous commit, it assumed format is specified and
not compressed case.

In this commit, it supports compressed data, too.

NOTE: In the test case, the boundary value which cause the overflow is
calculated in the advance because there is a case that the value
varies on machine.

See https://github.com/fluent/fluentd/pull/3553#issuecomment-972552443

Signed-off-by: Kentaro Hayashi <hayashi@clear-code.com>