Enable server plugins to specify socket-option `SO_LINGER`
`in_forward` plugin has the config, but most of plugins using the `server`
helper doesn't have the config, such as `in_tcp` or `in_http`.

The helper sets `0` to this option by default.
It causes the server sends RST rather than FIN on closing.
So this value needs to be changed if needing FIN signal on closing.

Adding this option to `transport_config` enable all plugins using the helper
to change this value.

Signed-off-by: daipom <fukuda@clear-code.com>