Change sigdump path on Windows
To "%windir%/Temp/fluentd-sigdump-#{Process.pid}.log" by default.

When 'SIGDUMP_PATH' environment variable is specified,
add pid to the extension of the path
in order to avoid a conflict of writing to the same file.

Made no modifications for UNIX-like for this commit.
As for UNIX-like, in order to add the same specification, 
need to add a trap of SIGCONT to the worker processes
or to fix the Sigdump library.
However, it loses backward compatibility for users using 'SIGDUMP_PATH'
environment variable, so need to consider what to do.

Signed-off-by: Daijiro Fukuda <fukuda@clear-code.com>