Abolish SIGCONT trap on UNIX-like
Shouldn't change SIGCONT behavior on UNIX-like for backward compatibility.
The rpc dump command is a new function, so make it send SIGCONT to all fluentd
processes.

With `SIGDUMP_PATH` environment variable specified on UNIX-like,
the rpc dump command will cause a writing conflict.
This should be fixed by improving the Sigdump library so that we can
specify a directory path.

Signed-off-by: Daijiro Fukuda <fukuda@clear-code.com>