in_tail: Show more information on skipping update_watcher
In #3614 some users reported that in_tail rarely stops tailing. It seems
it's caused when skipping update_watcher due to unexpected duplicate
pos_file entries. The root cause is unknown yet. To investigate it, show
more infomation when it happens. Ideally it should't happen so that log
level "warn" is desired for it.

Signed-off-by: Takuro Ashie <ashie@clear-code.com>