out_forward: Make sure to run ack thread on Ruby 3.2 on Windows
Override Coolio::Server#on_connection to protect local `evloop`
variable from GC. It looks like it can be freed by GC from time to
time, but I'm not sure it yet. So I've added the fix into our code
for testing instead of fixing upstream.

We'll report it to upstream after we verify the fix.

Signed-off-by: Takuro Ashie <ashie@clear-code.com>