Refactor logger initialization
In the supervisor process, Fluentd initializes the logger twice,
at `Supervisor::configure()` and `Supervisor.load_config()`.

`Supervisor.load_config()` is called in ServerEngine's reloading
function, but Fluentd doesn't use the function even when `SIGHUP`
or `SIGUSR2`.
So I can't see the reason for initializing the logger in the
callback.

In addition, I removed some needless parameters that were passed
to ServerEngine.

This also fixes a problem that `ignore_same_log_interval` is not
passed to the `Supervisor.load_config()` and that option is not
reflected to the logger of the supervisor process.

Signed-off-by: Daijiro Fukuda <fukuda@clear-code.com>