in_tail: prevent wrongly unwatching with follow_inodes
We must not unwatch targets that still exist.
If unwatching an existing target, it causes log duplication.

The mechanism is as follows.

The TailWatcher for rotated old file can be newly created
because the path is changed (the inode is the same).
(file.log -> file.log.1)

If unwatching the inode before creating that new TailWatcher,
it can't refer to the previous position and it reads logs in
duplicate.

Signed-off-by: Daijiro Fukuda <fukuda@clear-code.com>