Fix ChunkMessagePackEventStreamer's invalid argument
The argument `unpacker` of `ChunkMessagePackEventStreamer.each`
seems to have been added in order to match the feature with
`EventStream` at c6c6c038c5cacab6fcb9aa89a714d9366ac4902e.

However, that previous implementation at that point does not
work as expected.
It has never causes any issues just because the argument was not
used at all.

It could be implemented so that this argument is used, but given
that it has not been used so far, there is little need for it.

Signed-off-by: Daijiro Fukuda <fukuda@clear-code.com>