in_tail: use inode for key of TailWatchers when follow_inodes
This may improve the maintainability.

Note: Regardless of `follow_inodes`, both path or inode is acceptable
for the key of `tails`.
It is not wrong that the current logic uses `path` for the key.

Perhaps we don't need to use different keys depending on
`follow_inodes`. We can always use inode for the key. (Future work)

Signed-off-by: Katuya Kawakami <kkawakami-je@nec.com>
Signed-off-by: Masaki Hatada <mas-hatada@nec.com>
Co-authored-by: Daijiro Fukuda <fukuda@clear-code.com>