Add parameter to skip ignoring files
In our use-case we have a race condition.

We are creating files, then we are setting up privileges.
In function 'extend_paths' we can see that if file is not readable once it is ignored, so it happened to us, that some logs were ignored forever.

I think that introducing such parameter should be the easiest way to avoid similar problems in the future.

Signed-off-by: Artur Laskowski <apilaskowski@google.com>