test: fix timecop version to keep clock specs in tests
Backported from 57f821b194516fabd8b8d396842932126c6cbaf5.

---

timecop 0.9.9 supports `Process.clock_gettime`.
This breaks specifications of `process_extenstion` of Fluentd and
`Fluent::Clock`.

`Fluent::Clock` uses `CLOCK_MONOTONIC_RAW` if possible and it does not
be affected.
However, `CLOCK_MONOTONIC_RAW` is not available on Windows, so the
impact on tests on Windows is very significant.

For now, we should avoid this effect by fixing the version.

Signed-off-by: Daijiro Fukuda <fukuda@clear-code.com>