chumask: fix issue where it only works with standalone worker (#4836)
**Which issue(s) this PR fixes**: 
* Fixes #3671

**What this PR does / why we need it**: 
Fix #3671 (88a72608f36e8dd7287b44f1908f74fe19e5034a)

After #3671, the value of `chumask` passed to ServerEngine was always
`nil`.
The `chumask` value was being used only when standalone. It appears to
be unintentional.

This fixes it.

* By default, `0` as `chumask` is passed to ServerEngine.
* If specifying `--umask` option, that value is passed.
* This `chumask` value is applied when using `daemonize`.
  * When not using `daemonize`, the value is not used.
  * This is the specification of ServerEngine.

Specification change:
(Just undo the specification that #3671 unintentionally changed.)

* This changes the default `umask` when using `daemonize`.
  * fluent-package (RPM/DEB), for example.
  * Before: system default (`nil` for ServerEngine)
  * After: `0`

**Docs Changes**:
Not needed.
(There is not enough information on
https://docs.fluentd.org/deployment/command-line-option. It would be
good to add more information.)

**Release Note**: 
* Fixed a bug where the default `umask` was not set to `0` when using
`--daemon` (td-agent, fluent-package) since v1.14.6.
* `--umask` command line option: Fixed so that it is applied when
Fluentd runs with `--daemon` (fluent-package) as well as when Fluentd
runs with `--no-supervisor`.

Signed-off-by: Daijiro Fukuda <fukuda@clear-code.com>