Backport(v1.16): test_in_tail: add wait to detect rotation (#4874) (#4925)
**Which issue(s) this PR fixes**: 
Backport #4874
Fixes #4079

**What this PR does / why we need it**: 
For reasons such as the load on the CI execution environment, looks like
that rotation detection is delayed and the test does not execute as
expected.

For this reason, this patch adds a wait to ensure to detect the
rotation.

Before the patch was introduced, tests failed about once every 20 to 50
times.
After, the test was successful  500 consecutive times.

**Docs Changes**:

**Release Note**:

Signed-off-by: Shizuo Fujita <fujita@clear-code.com>
Signed-off-by: Kentaro Hayashi <hayashi@clear-code.com>
Co-authored-by: Shizuo Fujita <fujita@clear-code.com>