Add warning for directory permission to in_tail.rb (#4865)
**Which issue(s) this PR fixes**: 
~~Fixes #~~

**What this PR does / why we need it**: 
This PR provides a warning for typical configuration mistakes that
prevent log transfer from starting.

When using the tail type as a source, it is common for directories with
inappropriate partitions to be mistakenly included in the source path.

In my opinion, it would be preferable to issue a warning in such cases
to help users avoid these mistakes.

Steps to Reproduce:

```shell
$ sudo mkdir /var/log/noaccess
$ sudo mkdir /var/log/noaccess/noaccess2
$ sudo touch /var/log/noaccess/noaccess2/fluentd.log
$ sudo chmod 0000 /var/log/noaccess
$ sed -Ei 's|^([[:space:]]*)path .*|\1path /var/log/noaccess/noaccess2/fluentd.log|' example/in_tail.conf
$ bundle exec fluentd -c example/in_tail.conf
{no warning...}
```

**Docs Changes**:
No documentation changes required.

**Release Note**: 
in_tail: add warning for directory permission.

Signed-off-by: Tomoaki Kobayashi <tomoaki.kobayashi.t3@alumni.tohoku.ac.jp>