Limit high-memory logging ('pattern not matched' in parsing) to necessary log level (#4995)
**Which issue(s) this PR fixes**: 
Fixes #

**What this PR does / why we need it**: 
Currently, the “pattern not matched” warning is always generated
regardless of the `log_level` setting. This warning might use a lot of
memory when the log is very long.

To reduce memory usage, this PR will generate warnings only when
necessary, according to `log_level`.

Related to https://github.com/fluent/fluentd/issues/4994

* config
```
<system>
  log_level error
</system>

<source>
  @type tail
  tag test
  path "#{File.expand_path '~/tmp/fluentd/access-*.log'}"
  refresh_interval 5s
  read_from_head true
  <parse>
    @type regexp
    expression /^(?<log>a+)+$/
    timeout 0.1s
  </parse>
</source>

<match **>
  @type stdout
</match>
```

Here is memory usage.

![](https://github.com/user-attachments/assets/613ed741-a285-4e21-99f2-acdc88fbab13)



**Docs Changes**:
Not needed.

**Release Note**: 
Same as the title.

Signed-off-by: Shizuo Fujita <fujita@clear-code.com>