out_file: add `symlink_path_use_relative` option to use relative path instead of absolute path in `symlink_path` (#4904)
**Which issue(s) this PR fixes**: 
None.

**What this PR does / why we need it**: 
This PR moves from an absolute link to a relative link for
`symlink_path` parameter. Otherwise, this does not work in a dockerized
environment where the path is wired to the the outside of the container.
I can also add an additional parameter to switch between absolute and
relative linking if necessary and/or desired.

**Docs Changes**:
TODO https://docs.fluentd.org/input/tail

**Release Note**: 
Same as the title.

---------

Signed-off-by: Florian Dobener <florian.dobener@dwd.de>
Signed-off-by: domna <florian.dobener@physik.hu-berlin.de>