in_tail: fix handling of `encoding` parameter (#5010)
**Which issue(s) this PR fixes**: 
Fixes #4863

**What this PR does / why we need it**: 
If it configures `encoding` only (without `from_encoding`), it will
convert the string encoding unexpectedly and it might break the string.

This PR will fix this behavior and we can handle the string correctly.
As a result, the behavior will be aligned with the documented in
https://docs.fluentd.org/input/tail#encoding-from_encoding .

**Docs Changes**:
Not needed.

**Release Note**: 
in_tail: fixed where specifying only `encoding` parameter might cause
data corruption (affects since v0.14.12).

---------

Signed-off-by: Shizuo Fujita <fujita@clear-code.com>
Signed-off-by: Daijiro Fukuda <fukuda@clear-code.com>
Co-authored-by: Daijiro Fukuda <fukuda@clear-code.com>