in_http: replace WEBrick::HTTPUtils.parse_query with URI (#4900)
**Which issue(s) this PR fixes**: 
* Partially Fixes #4648

**What this PR does / why we need it**: 
Replace `WEBrick::HTTPUtils.parse_query` with `URI.decode_www_form` in
`in_http`.

WEBrick is no longer recommended for production use.

`WEBrick::HTTPUtils` is just a util, so it might be harmless.
However, it would be happy to reduce dependence if possible.

This changes some specifications unintentionally.
At least, this makes us unable to use `;` delimitter.
We need to consider whether this breaking change is acceptable or not.

**Docs Changes**:
Not needed.

**Release Note**: 
The same as the title.

Signed-off-by: Daijiro Fukuda <fukuda@clear-code.com>