out_stdout: support output to $stdout independently of logger (#5036)
**Which issue(s) this PR fixes**: 
* Fixes #5029

**What this PR does / why we need it**: 
Add `use_logger` option to `out_stdout` plugin.

* Type: `bool`
* Default: `true`

If you set `false` to this option, `out_stdout` outputs events to
`$stdout`, instead of `$logger`.
This feature enables events to be output to STDOUT in containerized
environments, such as K8s.

This does not change the default behavior.

Note: Although it's preferable for the default value of a `bool` option
to be `false` (because it can be used as a flag), I couldn't think of a
suitable name for it.

**Docs Changes**:
https://github.com/fluent/fluentd-docs-gitbook/pull/593

**Release Note**: 
out_stdout: support output to STDOUT independently of Fluentd logger by
setting `use_logger` to `false`.

---------

Signed-off-by: Daijiro Fukuda <fukuda@clear-code.com>
Co-authored-by: Takuro Ashie <ashie@clear-code.com>
Co-authored-by: Kentaro Hayashi <hayashi@clear-code.com>
Co-authored-by: Shizuo Fujita <fujita@clear-code.com>
Co-authored-by: Kentaro Hayashi <kenhys@gmail.com>